Add tests for existing Profile toString formats
Currently the Profile and LightweightProfile interfaces are independent
despite similar functionality. This adds tests for the SimpleProfile and MultipleSequenceAlignment, mostly to document existing formats. There
are some substantial differences in some formats between the two classes
which it might make sense to standardize.

Tests are included as resource files so they are easier to track in git.

SimpleProfileTest has been updated to junit 5 and fixed the argument
order, resulting in a big diff. The only major change should be the
addition of testToStringFormattedFiles.